I FIXED IT
I mean, the syscall still doesn't work, but at least there's no longer a party in my screen

+++ Kernel/kernel.c
@@ -88,7 +88,7 @@ void *initializeKernelBinary()
 	clearIDT();
 
 	ncPrint("  syscall: 0x");
-	ncPrintHex(idtEntry(0x80, syscall_hi));
+	ncPrintHex((uint64_t)idtEntry(0x80, syscall_hi));
 	ncNewline();
 
 	ncPrint("[Done]");
